+++ components/gitpod-protocol/src/frontend-dashboard-service.ts
@@ -51,7 +51,6 @@ export namespace IDEFrontendDashboardService {
         workspaceDescription: string;
         workspaceType: string;
         credentialsToken: string;
-        contextUrl: string;
     }
 
     export interface SetStateData {
