+++ components/supervisor/frontend/package.json
@@ -12,8 +12,7 @@
     "stream-browserify": "^2.0.1",
     "url": "^0.11.1",
     "util": "^0.11.1",
-    "uuid": "8.3.2",
-    "configcat-js": "^6.0.0"
+    "uuid": "8.3.2"
   },
   "devDependencies": {
     "@types/sharedworker": "^0.0.29",
